% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{package}
\name{BEDCA-package}
\alias{BEDCA}
\alias{BEDCA-package}
\title{BEDCA: Basic Engineering Data Collection and Analysis}
\description{
Data sets and example analyses from "Basic Engineering Data Collection and Analysis" by Stephen B. Vardeman and J. Marcus Jobe (2001, Brooks/Cole), republished 2023 by Iowa State University Digital Press.
}
\author{
\strong{Maintainer}: Michael Laviolette \email{statman54@gmail.com}

}
\keyword{internal}
