% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch1_eoc_prob4}
\alias{ch1_eoc_prob4}
\title{Chapter 1, Exercise 4}
\format{
A data frame with 8 rows and 3 variables.
 \describe{
  \item{Bushing}{Spanner bushing measured}
  \item{A}{Measurements by student technician A (inches)}
  \item{B}{Measurements by student technician B (inches)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 23
}
\usage{
data(ch1_eoc_prob4)
}
\description{
Chapter 1, Exercise 4
}
\examples{
data(ch1_eoc_prob4)
str(ch1_eoc_prob4)
}
\keyword{datasets}
