% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch1_xmp07}
\alias{ch1_xmp07}
\title{Chapter 1, Example 7}
\format{
A data frame with 10 rows and 2 variables.
 \describe{
  \item{tech}{Student who made measurement}
  \item{thickness}{Thickness of book paper (mm)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) Basic Engineering Data Collection and Analysis
\url{https://www.iastatedigitalpress.com/plugins/books/127/}
}
\usage{
data(ch1_xmp07)
}
\description{
Chapter 1, Example 7
}
\examples{
data(ch1_xmp07)
str(ch1_xmp07)

}
\keyword{datasets}
