% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch2_sec2_prob2}
\alias{ch2_sec2_prob2}
\title{Chapter 2, Section 2, Problem 2}
\format{
A data frame with 39 rows and 2 variables.
 \describe{
  \item{Laid}{Thrust face runouts for laid gears (0.0001 in.)}
  \item{Hung}{Thrust face runouts for hung gears (0.0001 in.)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) Basic Engineering Data Collection and Analysis
\url{https://www.iastatedigitalpress.com/plugins/books/127/}
}
\usage{
data(ch1_xmp07)
}
\description{
Chapter 2, Section 2, Problem 2
}
\examples{
data(ch2_sec2_prob2)
str(ch2_sec2_prob2)

}
\keyword{datasets}
