% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob01}
\alias{ch3_eoc_prob01}
\title{Chapter 3, Chapter Exercise 1}
\format{
A data frame with 120 rows and 1 variable.
 \describe{
  \item{Gains}{Gain measured on amplifier (dB)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 113
}
\usage{
data(ch3_eoc_prob01)
}
\description{
Chapter 3, Chapter Exercise 1
}
\examples{
data(ch3_eoc_prob01)
str(ch3_eoc_prob01)

}
\keyword{datasets}
