% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob02}
\alias{ch3_eoc_prob02}
\title{Chapter 3, Exercise 2}
\format{
A data frame with 26 rows and 1 variable.
 \describe{
  \item{ppm}{Amount of aluminum found in bihourly samples of PET (dB)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 113
}
\usage{
data(ch3_eoc_prob02)
}
\description{
Chapter 3, Exercise 2
}
\examples{
data(ch3_eoc_prob02)
str(ch3_eoc_prob02)

}
\keyword{datasets}
