% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob03}
\alias{ch3_eoc_prob03}
\title{Chapter 3, Exercise 3}
\format{
A data frame with 10 rows and 3 variables.
 \describe{
  \item{M1}{Measured manganese content in 1045 steel by method 1}
  \item{M2}{Measured manganese content in 1045 steel by method 2}
  \item{M3}{Measured manganese content in 1045 steel by method 3}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 114
}
\usage{
data(ch3_eoc_prob03)
}
\description{
Chapter 3, Exercise 3
}
\examples{
data(ch3_eoc_prob03)
str(ch3_eoc_prob03)

}
\keyword{datasets}
