% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob04}
\alias{ch3_eoc_prob04}
\title{Chapter 3, Chapter Exercise 4}
\format{
A data frame with 15 rows and 10 variables.
 \describe{
  \item{watts}{Wattage rating of resistor (watts)}
  \item{nominal}{Nominal resistance resistor (ohm)}
  \item{ohms}{Measured resistance of resistor (ohm)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 114
}
\usage{
data(ch3_eoc_prob04)
}
\description{
Chapter 3, Chapter Exercise 4
}
\examples{
data(ch3_eoc_prob04)
str(ch3_eoc_prob04)

}
\keyword{datasets}
