% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob06}
\alias{ch3_eoc_prob06}
\title{Chapter 3, Exercise 6}
\format{
A data frame with 12 rows and 1 variable.
 \describe{
  \item{life}{Tool lifetime, in holes drilled before failure}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 115
}
\usage{
data(ch3_eoc_prob06)
}
\description{
Chapter 3, Exercise 6
}
\note{
Data also used in Chapter 5, Exercise 22
}
\examples{
data(ch3_eoc_prob06)
str(ch3_eoc_prob06)

}
\keyword{datasets}
