% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob07}
\alias{ch3_eoc_prob07}
\title{Chapter 3, Exercise 7}
\format{
A data frame with 8 rows and 2 variables.
 \describe{
  \item{A}{Outside diameter of spanner bushing measured by Student A}
  \item{B}{Outside diameter of spanner bushing measured by Student B}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 116
}
\usage{
data(ch3_eoc_prob07)
}
\description{
Chapter 3, Exercise 7
}
\examples{
data(ch3_eoc_prob07)
str(ch3_eoc_prob07)

}
\keyword{datasets}
