% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob08}
\alias{ch3_eoc_prob08}
\title{Chapter 3, Chapter Exercise 8}
\format{
A data frame with 10 rows and 2 variables.
 \describe{
  \item{C1}{Time to failure of high-speed turbo engine bearing made from Compound 1}
  \item{C2}{Time to failure of high-speed turbo engine bearing made from Compound 2}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 116
}
\usage{
data(ch3_eoc_prob08)
}
\description{
Chapter 3, Chapter Exercise 8
}
\examples{
data(ch3_eoc_prob08)
str(ch3_eoc_prob08)

}
\keyword{datasets}
