% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob09}
\alias{ch3_eoc_prob09}
\title{Chapter 3, Exercise 9}
\format{
A data frame with 405 rows and 1 variable.
 \describe{
  \item{Height}{Punch height (0.001 in.)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 116
}
\usage{
data(ch3_eoc_prob09)
}
\description{
Chapter 3, Exercise 9
}
\examples{
data(ch3_eoc_prob09)
str(ch3_eoc_prob09)

}
\keyword{datasets}
