% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob10}
\alias{ch3_eoc_prob10}
\title{Chapter 3, Chapter Exercise 10}
\format{
A data frame with 100 rows and 1 variable.
 \describe{
  \item{Purities}{Daily purities of oxygen, in increments of 0.01 percent above 99.00}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 117
}
\usage{
data(ch3_eoc_prob10)
}
\description{
Chapter 3, Chapter Exercise 10
}
\examples{
data(ch3_eoc_prob10)
str(ch3_eoc_prob10)

}
\keyword{datasets}
