% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob11}
\alias{ch3_eoc_prob11}
\title{Chapter 3, Exercise 11}
\format{
A data frame with 10 rows and 4 variables.
 \describe{
  \item{Stres87}{Failure time for hardened steel specimens subjected to 0.87 x 10^6 psi}
  \item{Stres99}{Failure time for hardened steel specimens subjected to 0.99 x 10^6 psi}
  \item{Stres109}{Failure time for hardened steel specimens subjected to 1.09 x 10^6 psi}
  \item{Stres118}{Failure time for hardened steel specimens subjected to 1.18 x 10^6 psi}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 117
}
\usage{
data(ch3_eoc_prob11)
}
\description{
Chapter 3, Exercise 11
}
\note{
Time units are unknown
}
\examples{
data(ch3_eoc_prob11)
str(ch3_eoc_prob11)

}
\keyword{datasets}
