% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob12}
\alias{ch3_eoc_prob12}
\title{Chapter 3, Chapter Exercise 12}
\format{
A data frame with 20 rows and 3 variables.
 \describe{
  \item{Inch}{Nominal length of steel strip}
  \item{X14G}{Measured length of 14 gauge steel strip (0.001 in. above nominal)}
  \item{X16G}{Measured length of 16 gauge steel strip (0.001 in. above nominal)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 118
}
\usage{
data(ch3_eoc_prob12)
}
\description{
Chapter 3, Chapter Exercise 12
}
\examples{
data(ch3_eoc_prob12)
str(ch3_eoc_prob12)

}
\keyword{datasets}
