% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob13}
\alias{ch3_eoc_prob13}
\title{Chapter 3, Chapter Exercise 13}
\format{
A data frame with 12 rows and 1 variable.
 \describe{
  \item{Resist}{Resistivity of doped polysilicon specimens}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 118
}
\usage{
data(ch3_eoc_prob13)
}
\description{
Chapter 3, Chapter Exercise 13
}
\note{
Units unknown
}
\examples{
data(ch3_eoc_prob13)
str(ch3_eoc_prob13)

}
\keyword{datasets}
