% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob14}
\alias{ch3_eoc_prob14}
\title{Chapter 3, Exercise 14}
\format{
A data frame with 10 rows and 2 variables.
 \describe{
  \item{Lab1}{Lifetime of wire specimens (in weeks) measured by Laboratory 1}
  \item{Lab2}{Lifetime of wire specimens (in weeks) measured by Laboratory 2}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 118
}
\usage{
data(ch3_eoc_prob14)
}
\description{
Chapter 3, Exercise 14
}
\examples{
data(ch3_eoc_prob14)
str(ch3_eoc_prob14)

}
\keyword{datasets}
