% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob15}
\alias{ch3_eoc_prob15}
\title{Chapter 3, Chapter Exercise 15}
\format{
A data frame with 12 rows and 3 variables.
 \describe{
  \item{Board}{Specimen of 4 ft x 8 ft wallboard}
  \item{Before}{Thickness (in inches) before drying in a kiln}
  \item{After}{Thickness (in inches) after drying in a kiln}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 119
}
\usage{
data(ch3_eoc_prob15)
}
\description{
Chapter 3, Chapter Exercise 15
}
\examples{
data(ch3_eoc_prob15)
str(ch3_eoc_prob15)

}
\keyword{datasets}
