% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob16}
\alias{ch3_eoc_prob16}
\title{Chapter 3, Exercise 16}
\format{
A data frame with 20 rows and 1 variable.
 \describe{
  \item{Diameter}{Particle diameter observed in micrograph (10^-8 m)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 119
}
\usage{
data(ch3_eoc_prob16)
}
\description{
Chapter 3, Exercise 16
}
\examples{
data(ch3_eoc_prob16)
str(ch3_eoc_prob16)

}
\keyword{datasets}
