% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob17}
\alias{ch3_eoc_prob17}
\title{Chapter 3, Chapter Exercise 17}
\format{
A data frame with 13 rows and 2 variables.
 \describe{
  \item{A}{Latent heat of fusion of ice, measured by electrical method}
  \item{B}{Latent heat of fusion of ice, measured by mixture of methods}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 120
}
\usage{
data(ch3_eoc_prob17)
}
\description{
Chapter 3, Chapter Exercise 17
}
\examples{
data(ch3_eoc_prob17)
str(ch3_eoc_prob17)

}
\keyword{datasets}
