% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob18}
\alias{ch3_eoc_prob18}
\title{Chapter 3, Exercise 18}
\format{
A data frame with 10 rows and 3 variables.
 \describe{
  \item{H1}{Lives until failure after testing on rotary fatigue strength machine, Heat 1}
  \item{H2}{Lives until failure after testing on rotary fatigue strength machine, Heat 2}
  \item{H3}{Lives until failure after testing on rotary fatigue strength machine, Heat 3}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 120
}
\usage{
data(ch3_eoc_prob18)
}
\description{
Chapter 3, Exercise 18
}
\examples{
data(ch3_eoc_prob18)
str(ch3_eoc_prob18)

}
\keyword{datasets}
