% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob19}
\alias{ch3_eoc_prob19}
\title{Chapter 3, Exercise 19}
\format{
A data frame with 10 rows and 4 variables.
 \describe{
  \item{Sleeve}{Sleeve identifier}
  \item{N.DB}{Outside diameter measurement, notch-dial bore}
  \item{NN.DB}{Outside diameter measurement, non-notch-dial bore}
  \item{N.AS}{Outside diameter measurement, notch-air splinder}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 120
}
\usage{
data(ch3_eoc_prob19)
}
\description{
Chapter 3, Exercise 19
}
\note{
Units are 10^-5 inches above nominal. Data also used in Chapter 6, Exercise 33
}
\examples{
data(ch3_eoc_prob19)
str(ch3_eoc_prob19)

}
\keyword{datasets}
