% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob20}
\alias{ch3_eoc_prob20}
\title{Chapter 3, Exercise 20}
\format{
A data frame with 13 rows and 3 variables.
 \describe{
  \item{LaserA}{Angle of laser-drilled hole to top surface of part, Hole A}
  \item{EDM}{Angle of hole drilled by electrical discharge machining to top surface of part, Hole A}
  \item{LaserB}{Angle of laser-drilled hole to top surface of part, Hole B}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 121
}
\usage{
data(ch3_eoc_prob20)
}
\description{
Chapter 3, Exercise 20
}
\note{
Units are degrees; measurements for EDM Hole B not given.
}
\examples{
data(ch3_eoc_prob20)
str(ch3_eoc_prob20)

}
\keyword{datasets}
