% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob21}
\alias{ch3_eoc_prob21}
\title{Chapter 3, Exercise 21}
\format{
A data frame with 10 rows and 3 variables.
 \describe{
  \item{Dial}{Brinell hardness, measured with dial Rockwell tester}
  \item{Digital}{Brinell hardness, measured with digital Rockwell tester}
  \item{Brinell}{Brinell hardness, measured with Brinell tester}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 122
}
\usage{
data(ch3_eoc_prob21)
}
\description{
Chapter 3, Exercise 21
}
\examples{
data(ch3_eoc_prob21)
str(ch3_eoc_prob21)

}
\keyword{datasets}
