% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_eoc_prob22}
\alias{ch3_eoc_prob22}
\title{Chapter 3, Exercise 22}
\format{
A data frame with 10 rows and 2 variables.
 \describe{
  \item{D1}{Flight time for paper helicopter (seconds), design 1}
  \item{D2}{Flight time for paper helicopter (seconds), design 2}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 122
}
\usage{
data(ch3_eoc_prob22)
}
\description{
Chapter 3, Exercise 22
}
\examples{
data(ch3_eoc_prob22)
str(ch3_eoc_prob22)

}
\keyword{datasets}
