% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_sec1_prob1}
\alias{ch3_sec1_prob1}
\title{Section 3.1, Exercise 1}
\format{
A data frame with 40 rows and 1 variable.
 \describe{
  \item{runs}{Percent yields from 40 runs of a chemical process}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 77
}
\usage{
data(ch3_sec1_prob1)
}
\description{
Section 3.1, Exercise 1
}
\note{
Data also used in Section 3.2, Exercise 3
}
\examples{
data(ch3_sec1_prob1)
str(ch3_sec1_prob1)

}
\keyword{datasets}
