% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_sec1_prob2}
\alias{ch3_sec1_prob2}
\title{Chapter 3, Section 1, Problem 2}
\format{
A data frame with 20 rows and 2 variables.
 \describe{
  \item{230_G}{Bullet penetration depths for 230 grain jacketed bullets}
  \item{200_G}{Bullet penetration depths for 200 grain jacketed bullets}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 77
}
\usage{
data(ch3_sec1_prob2)
}
\description{
Chapter 3, Section 1, Problem 2
}
\note{
First appearance as Table 3.1, p. 67
}
\examples{
data(ch3_sec1_prob2)
str(ch3_sec1_prob2)

}
\keyword{datasets}
