% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_sec1_prob3}
\alias{ch3_sec1_prob3}
\title{Section 3.1, Exercise 3}
\format{
A data frame with 20 rows and 2 variables.
 \describe{
  \item{Piece}{Piece ID}
  \item{Top}{Torque required to loosen top bolt (ft-lb)}
  \item{Bottom}{Torque required to loosen bottom bolt (ft-lb)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 77
}
\usage{
data(ch3_sec1_prob3)
}
\description{
Section 3.1, Exercise 3
}
\examples{
data(ch3_sec1_prob3)
str(ch3_sec1_prob3)

}
\keyword{datasets}
