% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch3_sec2_prob1}
\alias{ch3_sec2_prob1}
\title{Section 3.2, Exercise 1}
\format{
A data frame with 10 rows and 2 variables.
 \describe{
  \item{Length}{Impact strength of insulating material cut lengthwise (ft-lb)}
  \item{Cross}{Impact strength of insulating material cut crosswise (ft-lb)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 92
}
\usage{
data(ch3_sec2_prob1)
}
\description{
Section 3.2, Exercise 1
}
\note{
Data also used in Section 3.3, Exercises 1 and 3; Section 6.3, Exercise 2
}
\examples{
data(ch3_sec2_prob1)
str(ch3_sec2_prob1)

}
\keyword{datasets}
