% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob01}
\alias{ch4_eoc_prob01}
\title{Chapter 4, Exercise 1}
\format{
A data frame with 9 rows and 2 variables.
 \describe{
  \item{x}{Water/cement ratio}
  \item{y}{14-day compressive strength (psi)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 203
}
\usage{
data(ch4_eoc_prob01)
}
\description{
Chapter 4, Exercise 1
}
\note{
Data also used in Section 9.1, Exercise 2; Chapter 9, Exercise 4(b)
}
\examples{
data(ch4_eoc_prob01)
str(ch4_eoc_prob01)

}
\keyword{datasets}
