% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob02}
\alias{ch4_eoc_prob02}
\title{Chapter 4, Exercise 2}
\format{
A data frame with 6 rows and 2 variables.
 \describe{
  \item{x}{Carburetor jetting size}
  \item{y}{Elapsed time (seconds)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 204
}
\usage{
data(ch4_eoc_prob02)
}
\description{
Chapter 4, Exercise 2
}
\note{
Data reused in Section 9.2, Exercise 1
}
\examples{
data(ch4_eoc_prob02)
str(ch4_eoc_prob02)

}
\keyword{datasets}
