% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob03}
\alias{ch4_eoc_prob03}
\title{Chapter 4, Exercise 3}
\format{
A data frame with 9 rows and 3 variables.
 \describe{
  \item{x1}{Temperature for heat treating, K}
  \item{x2}{Time for heat treating (minutes)}
  \item{y}{Mean grain diameter (micrometers)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 204
}
\usage{
data(ch4_eoc_prob03)
}
\description{
Chapter 4, Exercise 3
}
\examples{
data(ch4_eoc_prob03)
str(ch4_eoc_prob03)

}
\keyword{datasets}
