% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob05}
\alias{ch4_eoc_prob05}
\title{Chapter 4, Exercise 5}
\format{
A data frame with 8 rows and 5 variables.
 \describe{
  \item{A}{Diameter of mechanical pencil lead, coded -1 and 1}
  \item{B}{Length of lead protruding over desk, coded -1 and 1}
  \item{C}{Hardness (B or 2H), coded -1 and 1}
  \item{y1}{Number of paper clips loaded until failure, replication 1}
  \item{y2}{Number of paper clips loaded until failure, replication 2}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 205
}
\usage{
data(ch4_eoc_prob05)
}
\description{
Chapter 4, Exercise 5
}
\examples{
data(ch4_eoc_prob05)
str(ch4_eoc_prob05)

}
\keyword{datasets}
