% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob06}
\alias{ch4_eoc_prob06}
\title{Chapter 4, Exercise 6}
\format{
A data frame with 8 rows and 4 variables.
 \describe{
  \item{A}{Mesh size of powder particles, coded -1 and 1}
  \item{B}{Volume of graduated cylinder, coded -1 and 1}
  \item{C}{Vibration of cylinder, coded -1 and 1}
  \item{Y}{Mean density of crushed tabular alumina powder (g/cc)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 206
}
\usage{
data(ch4_eoc_prob06)
}
\description{
Chapter 4, Exercise 6
}
\examples{
data(ch4_eoc_prob06)
str(ch4_eoc_prob06)

}
\keyword{datasets}
