% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob07}
\alias{ch4_eoc_prob07}
\title{Chapter 4, Exercise 7}
\format{
A data frame with 27 rows and 4 variables.
 \describe{
  \item{y}{Number of cycles until failure of worsted yarn}
  \item{x1}{Length (mm)}
  \item{x2}{Amplitude of loading cycle (mm)}
  \item{x3}{Load (g)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 206
}
\usage{
data(ch4_eoc_prob07)
}
\description{
Chapter 4, Exercise 7
}
\examples{
data(ch4_eoc_prob07)
str(ch4_eoc_prob07)

}
\keyword{datasets}
