% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob08}
\alias{ch4_eoc_prob08}
\title{Chapter 4, Exercise 8}
\format{
A data frame with 30 rows and 3 variables.
 \describe{
  \item{Size}{Size of charge (2.5 mL, 5.0 mL)}
  \item{Prop}{Type of propellant (lighter fluid, gasoline, carburetor fluid)}
  \item{Distance}{Distance traveled by tennis ball shot from Polish cannon (feet)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 207
}
\usage{
data(ch4_eoc_prob08)
}
\description{
Chapter 4, Exercise 8
}
\examples{
data(ch4_eoc_prob08)
str(ch4_eoc_prob08)

}
\keyword{datasets}
