% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob09}
\alias{ch4_eoc_prob09}
\title{Chapter 4, Exercise 9}
\format{
A data frame with 16 rows and 5 variables.
 \describe{
  \item{A}{Ceramic surface: unglazed (-1) vs. glazed (+1)}
  \item{B}{Metal film thickness: normal (-1) vs. 1.5 times normal (+1)}
  \item{C}{Annealing Time: normal (-1) vs. 4 times normal (+1)}
  \item{D}{Prebond clean: normal clean (-1) vs. no clean (+1)}
  \item{Pullouts}{Number of pullouts}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 208
}
\usage{
data(ch4_eoc_prob09)
}
\description{
Chapter 4, Exercise 9
}
\note{
Data reused in Section 8.2, Exercise 2 and Chapter 8, Exercise 9
}
\examples{
data(ch4_eoc_prob09)
str(ch4_eoc_prob09)

}
\keyword{datasets}
