% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob12}
\alias{ch4_eoc_prob12}
\title{Chapter 4, Exercise 12}
\format{
A data frame with 36 rows and 3 variables.
 \describe{
  \item{x1}{Cutting speed (sfpm)}
  \item{x2}{Feed (ipr)}
  \item{y}{Tool life (minutes)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 209
}
\usage{
data(ch4_eoc_prob12)
}
\description{
Chapter 4, Exercise 12
}
\examples{
data(ch4_eoc_prob12)
str(ch4_eoc_prob12)

}
\keyword{datasets}
