% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob13}
\alias{ch4_eoc_prob13}
\title{Chapter 4, Exercise 13}
\format{
A data frame with 8 rows and 5 variables.
 \describe{
  \item{Test}{Test sequence}
  \item{mph}{Speed (mph)}
  \item{Octane}{Gasoline octane}
  \item{Used}{Gallon of gasoline used}
  \item{mpg}{Mileage (mpg)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 209
}
\usage{
data(ch4_eoc_prob13)
}
\description{
Chapter 4, Exercise 13
}
\examples{
data(ch4_eoc_prob13)
str(ch4_eoc_prob13)

}
\keyword{datasets}
