% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob14}
\alias{ch4_eoc_prob14}
\title{Chapter 4, Exercise 14}
\format{
A data frame with 27 rows and 3 variables.
 \describe{
  \item{x1}{Speed (rpm)}
  \item{x2}{Load (psi)}
  \item{y}{Wear life (hours)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 210
}
\usage{
data(ch4_eoc_prob14)
}
\description{
Chapter 4, Exercise 14
}
\examples{
data(ch4_eoc_prob14)
str(ch4_eoc_prob14)

}
\keyword{datasets}
