% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob15}
\alias{ch4_eoc_prob15}
\title{Chapter 4, Exercise 15}
\format{
A data frame with 27 rows and 3 variables.
 \describe{
  \item{y1}{mL of 1N H2SO4 used in the first phase of titration}
  \item{y2}{mL of 1N H2SO4 used in the second phase of titration}
  \item{x1}{Percent by volume of component A}
  \item{x2}{Percent by volume of component B}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 211
}
\usage{
data(ch4_eoc_prob15)
}
\description{
Chapter 4, Exercise 15
}
\examples{
data(ch4_eoc_prob15)
str(ch4_eoc_prob15)

}
\keyword{datasets}
