% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob16}
\alias{ch4_eoc_prob16}
\title{Chapter 4, Exercise 16}
\format{
A data frame with 10 rows and 3 variables.
 \describe{
  \item{Batch}{Batch ID}
  \item{y}{Compressive strength (psi)}
  \item{x}{Splitting tensile strength (psi)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 211
}
\usage{
data(ch4_eoc_prob16)
}
\description{
Chapter 4, Exercise 16
}
\note{
Data also used in Chapter 9, Exercise 6
}
\examples{
data(ch4_eoc_prob16)
str(ch4_eoc_prob16)

}
\keyword{datasets}
