% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob18}
\alias{ch4_eoc_prob18}
\title{Chapter 4, Exercise 18}
\format{
A data frame with 9 rows and 6 variables.
 \describe{
  \item{x1}{Agitator speed (rpm)}
  \item{x2}{Polymer concentration (ppm)}
  \item{y1}{Percent recovery of pyrite}
  \item{y2}{Percent recover of kaolin}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 212
}
\usage{
data(ch4_eoc_prob18)
}
\description{
Chapter 4, Exercise 18
}
\note{
Data also used in Chapter 4, Exercise 19
}
\examples{
data(ch4_eoc_prob18)
str(ch4_eoc_prob18)

}
\keyword{datasets}
