% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob20}
\alias{ch4_eoc_prob20}
\title{Chapter 4, Exercise 20}
\format{
A data frame with 16 rows and 5 variables.
 \describe{
  \item{A}{Plane design: straight wing (-1) vs. t wing (+1)}
  \item{B}{Nose weight: none (-1) vs. paper clip (+1)}
  \item{C}{Paper hype: notebook (-1) vs. construction (+1)}
  \item{D}{Wing tips: straight (-1) vs. bent up (+1)}
  \item{y}{Flight distance of paper airplane}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 213
}
\usage{
data(ch4_eoc_prob20)
}
\description{
Chapter 4, Exercise 20
}
\examples{
data(ch4_eoc_prob20)
str(ch4_eoc_prob20)

}
\keyword{datasets}
