% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob21}
\alias{ch4_eoc_prob21}
\title{Chapter 4, Exercise 21}
\format{
A data frame with 20 rows and 3 variables.
 \describe{
  \item{x1}{Thickness of armor plate (0.001 in.)}
  \item{x2}{Brinell hardness number}
  \item{y}{Ballistic limit (ft/sec)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 214
}
\usage{
data(ch4_eoc_prob21)
}
\description{
Chapter 4, Exercise 21
}
\examples{
data(ch4_eoc_prob21)
str(ch4_eoc_prob21)

}
\keyword{datasets}
