% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob22}
\alias{ch4_eoc_prob22}
\title{Chapter 4, Exercise 22}
\format{
A data frame with 6 rows and 3 variables.
 \describe{
  \item{Width}{Width of bicycle tire}
  \item{Tread}{smooth (1), treaded (2)}
  \item{y}{Time to stop (sec)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 214
}
\usage{
data(ch4_eoc_prob22)
}
\description{
Chapter 4, Exercise 22
}
\examples{
data(ch4_eoc_prob22)
str(ch4_eoc_prob22)

}
\keyword{datasets}
