% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob23}
\alias{ch4_eoc_prob23}
\title{Chapter 4, Exercise 23}
\format{
A data frame with 10 rows and 2 variables.
 \describe{
  \item{x}{Density of PETN (g/cc)}
  \item{y}{Detonation velocity (km/sec)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 215
}
\usage{
data(ch4_eoc_prob23)
}
\description{
Chapter 4, Exercise 23
}
\examples{
data(ch4_eoc_prob23)
str(ch4_eoc_prob23)

}
\keyword{datasets}
