% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob24}
\alias{ch4_eoc_prob24}
\title{Chapter 4, Exercise 24}
\format{
A data frame with 19 rows and 2 variables.
 \describe{
  \item{x}{Depth of stud (0.001 in. above 0.400)}
  \item{y}{Torque at failure (lbs/in)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 215
}
\usage{
data(ch4_eoc_prob24)
}
\description{
Chapter 4, Exercise 24
}
\examples{
data(ch4_eoc_prob24)
str(ch4_eoc_prob24)

}
\keyword{datasets}
