% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob25}
\alias{ch4_eoc_prob25}
\title{Chapter 4, Exercise 25}
\format{
A data frame with 19 rows and 2 variables.
 \describe{
  \item{x}{Amount of drag applied to tire (in percent)}
  \item{y}{Grip force (y)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 216
}
\usage{
data(ch4_eoc_prob25)
}
\description{
Chapter 4, Exercise 25
}
\examples{
data(ch4_eoc_prob25)
str(ch4_eoc_prob25)

}
\keyword{datasets}
