% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob26}
\alias{ch4_eoc_prob26}
\title{Chapter 4, Exercise 26}
\format{
A data frame with 24 rows and 2 variables.
 \describe{
  \item{x}{Asphalt content (percent by weight)}
  \item{y}{Permeability (water loss in inches/hour)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 216
}
\usage{
data(ch4_eoc_prob26)
}
\description{
Chapter 4, Exercise 26
}
\note{
Data also used in Chapter 9, Exercise 13
}
\examples{
data(ch4_eoc_prob26)
str(ch4_eoc_prob26)

}
\keyword{datasets}
