% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob27}
\alias{ch4_eoc_prob27}
\title{Chapter 4, Exercise 27}
\format{
A data frame with 36 rows and 3 variables.
 \describe{
  \item{x1}{Diameter of wooden dowel rod (in.)}
  \item{x2}{Length of wooden dowel rod (in.)}
  \item{y}{Breaking strength (lb.)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 217
}
\usage{
data(ch4_eoc_prob27)
}
\description{
Chapter 4, Exercise 27
}
\examples{
data(ch4_eoc_prob27)
str(ch4_eoc_prob27)

}
\keyword{datasets}
