% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob28}
\alias{ch4_eoc_prob28}
\title{Chapter 4, Exercise 28}
\format{
A data frame with 12 rows and 4 variables.
 \describe{
  \item{x1}{Diameter of aluminum alloy (in.)}
  \item{x2}{Feed rate (in. rev)}
  \item{y1}{Thrust (lb.)}
  \item{y2}{Torque (ft-lb.)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 218
}
\usage{
data(ch4_eoc_prob28)
}
\description{
Chapter 4, Exercise 28
}
\examples{
data(ch4_eoc_prob28)
str(ch4_eoc_prob28)

}
\keyword{datasets}
