% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob29}
\alias{ch4_eoc_prob29}
\title{Chapter 4, Exercise 29}
\format{
A data frame with 16 rows and 5 variables.
 \describe{
  \item{A}{Amount of glue used}
  \item{B}{Predrying temperature}
  \item{C}{Tunnel temperature}
  \item{D}{Pressure applied}
  \item{y}{Adhesive force (N)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 219
}
\usage{
data(ch4_eoc_prob29)
}
\description{
Chapter 4, Exercise 29
}
\note{
Exact levels of variables not given.
}
\examples{
data(ch4_eoc_prob29)
str(ch4_eoc_prob29)

}
\keyword{datasets}
