% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob30}
\alias{ch4_eoc_prob30}
\title{Chapter 4, Exercise 30}
\format{
A data frame with 16 rows and 5 variables.
 \describe{
  \item{A}{Chamber pressure (torr): 8 (-1) vs. 9 (+1)}
  \item{B}{H2 flow (standard cm3/min): 500 (-1) vs. 1000 (+1)}
  \item{C}{SiH4 flow (standard cm3/min): 15 (-1) vs. 25 (+1)}
  \item{D}{WF6 Flow (standard cm3/min): 50 (-1) vs. 60 (+1)}
  \item{y}{Average resistivity of tungsten film (milliohms/cm)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 220
}
\usage{
data(ch4_eoc_prob30)
}
\description{
Chapter 4, Exercise 30
}
\examples{
data(ch4_eoc_prob30)
str(ch4_eoc_prob30)

}
\keyword{datasets}
