% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_sec1_prob1}
\alias{ch4_sec1_prob1}
\title{Chapter 4, Section 1, Exercise 1}
\format{
A data frame with 5 rows and 2 variables.
 \describe{
  \item{y}{Artificial data, y}
  \item{x}{Artificial data, x}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 139
}
\usage{
data(ch4_sec1_prob1)
}
\description{
Chapter 4, Section 1, Exercise 1
}
\note{
Data repeated in Chapter 4, Section 1, Exercise 2
}
\examples{
data(ch4_sec1_prob1)
str(ch4_sec1_prob1)

}
\keyword{datasets}
