% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_sec1_prob2}
\alias{ch4_sec1_prob2}
\title{Section 4.1, Exercise 1}
\format{
A data frame with 5 rows and 2 variables.
 \describe{
  \item{y}{Artificial data, y}
  \item{x}{Artificial data, x}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 139
}
\usage{
data(ch4_sec1_prob2)
}
\description{
Section 4.1, Exercise 1
}
\note{
Data repeated from Section 4.1, Exercise 1
}
\examples{
data(ch4_sec1_prob2)
str(ch4_sec1_prob2)

}
\keyword{datasets}
