% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_sec1_prob3}
\alias{ch4_sec1_prob3}
\title{Section 4.1, Exercise 3}
\format{
A data frame with 8 rows and 2 variables.
 \describe{
  \item{x}{Pot temperature (C)}
  \item{y}{Average molecular weight}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 140
}
\usage{
data(ch4_sec1_prob3)
}
\description{
Section 4.1, Exercise 3
}
\note{
Data repeated in Section 4.2, Exercise 1
}
\examples{
data(ch4_sec1_prob3)
str(ch4_sec1_prob3)

}
\keyword{datasets}
