% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_sec1_prob4}
\alias{ch4_sec1_prob4}
\title{Chapter 4, Section 1, Exercise 4}
\format{
A data frame with 20 rows and 2 variables.
 \describe{
  \item{y}{Tool life (minutes)}
  \item{x}{Cutting speed (sfpm)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 140
}
\usage{
data(ch4_sec1_prob4)
}
\description{
Chapter 4, Section 1, Exercise 4
}
\examples{
data(ch4_sec1_prob4)
str(ch4_sec1_prob4)

}
\keyword{datasets}
