% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_sec2_prob2}
\alias{ch4_sec2_prob2}
\title{Section 4.2, Exercise 2}
\format{
A data frame with 9 rows and 3 variables.
 \describe{
  \item{x1}{Percent NaOH}
  \item{x2}{Pretreatment time (minutes)}
  \item{y}{Specific surface area (cubic cm/g)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 162
}
\usage{
data(ch4_sec2_prob2)
}
\description{
Section 4.2, Exercise 2
}
\note{
Data repeated in Section 4.3, Exercise 1; Section 9.2, Exercise 2
}
\examples{
data(ch4_sec2_prob2)
str(ch4_sec2_prob2)

}
\keyword{datasets}
