% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_sec3_prob2}
\alias{ch4_sec3_prob2}
\title{Section 4.3, Exercise 2}
\format{
A data frame with 8 rows and 3 variables.
 \describe{
  \item{Tech}{Technician performing measurements, coded -1 and 1}
  \item{Diameter}{Diameter of 0.188 or 0.314 inches, coded -1 and 1}
  \item{Fluid}{Water or ethylene glycol, coded -1 and 1}
  \item{y1}{Time for fluid level to drop, replicate 1}
  \item{y2}{Time for fluid level to drop, replicate 2}
  \item{y3}{Time for fluid level to drop, replicate 3}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 191
}
\usage{
data(ch4_sec3_prob2)
}
\description{
Section 4.3, Exercise 2
}
\examples{
data(ch4_sec3_prob2)
str(ch4_sec3_prob2)

}
\keyword{datasets}
