% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch5_eoc_prob27}
\alias{ch5_eoc_prob27}
\title{Chapter 5, Exercise 27}
\format{
A data frame with 30 rows and 1 variable.
 \describe{
  \item{Strength}{Fracture strengths of specimens of hot-pressed silicon nitrides (MPa)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 328
}
\usage{
data(ch5_eoc_prob27)
}
\description{
Chapter 5, Exercise 27
}
\examples{
data(ch5_eoc_prob27)
str(ch5_eoc_prob27)

}
\keyword{datasets}
