% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch5_eoc_prob28}
\alias{ch5_eoc_prob28}
\title{Chapter 5, Exercise 28}
\format{
A data frame with 17 rows and 1 variable.
 \describe{
  \item{Life}{Number of eight-hour shifts before failure of sinker rollers}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 328
}
\usage{
data(ch5_eoc_prob28)
}
\description{
Chapter 5, Exercise 28
}
\examples{
data(ch5_eoc_prob28)
str(ch5_eoc_prob28)

}
\keyword{datasets}
