% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch5_eoc_prob29}
\alias{ch5_eoc_prob29}
\title{Chapter 5, Exercise 29}
\format{
A data frame with 24 rows and 1 variable.
 \describe{
  \item{distance}{Deviations from nominal of distance between two holes drilled in steel plates}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 329
}
\usage{
data(ch5_eoc_prob29)
}
\description{
Chapter 5, Exercise 29
}
\examples{
data(ch5_eoc_prob29)
str(ch5_eoc_prob29)

}
\keyword{datasets}
