% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch5_sec3_prob4}
\alias{ch5_sec3_prob4}
\title{Section 5.3, Exercise 4}
\format{
A data frame with 23 rows and 1 variable.
 \describe{
  \item{Bearings}{Lifetimes of ball bearings (10^6 revolutions before failure}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 278
}
\usage{
data(ch5_sec3_prob4)
}
\description{
Section 5.3, Exercise 4
}
\examples{
data(ch5_sec3_prob4)
str(ch5_sec3_prob4)

}
\keyword{datasets}
