% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch5_sec3_prob5}
\alias{ch5_sec3_prob5}
\title{Section 5.3, Exercise 5}
\format{
A data frame with 19 rows and 1 variable.
 \describe{
  \item{mileage}{Mileage at first failure for military personnel carriers}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 278
}
\usage{
data(ch5_sec3_prob5)
}
\description{
Section 5.3, Exercise 5
}
\examples{
data(ch5_sec3_prob5)
str(ch5_sec3_prob5)

}
\keyword{datasets}
