% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch5_xmp15}
\alias{ch5_xmp15}
\title{Chapter 5, Example 15}
\format{
A data frame with 65 rows and 1 variable.
 \describe{
  \item{sale_time}{Time (seconds) required to complete postage stamp sale at residence hall service counter}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 270
}
\usage{
data(ch5_xmp15)
}
\description{
Chapter 5, Example 15
}
\examples{
data(ch5_xmp15)
str(ch5_xmp15)

}
\keyword{datasets}
