% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch6_eoc_prob05}
\alias{ch6_eoc_prob05}
\title{Chapter 6, Exercise 5}
\format{
A data frame with 16 rows and 1 variable.
 \describe{
  \item{length}{Length of spanner bushings (in)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 428
}
\usage{
data(ch6_eoc_prob05)
}
\description{
Chapter 6, Exercise 5
}
\examples{
data(ch6_eoc_prob05)
str(ch6_eoc_prob05)

}
\keyword{datasets}
