% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch6_eoc_prob06}
\alias{ch6_eoc_prob06}
\title{Chapter 6, Exercise 6}
\format{
A data frame with 16 rows and 3 variables.
 \describe{
  \item{Bushing}{Spanner bushing ID}
  \item{A}{Outside diameter measured by Student A}
  \item{B}{Outside diameter measured by Student B}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 428
}
\usage{
data(ch6_eoc_prob06)
}
\description{
Chapter 6, Exercise 6
}
\examples{
data(ch6_eoc_prob06)
str(ch6_eoc_prob06)

}
\keyword{datasets}
