% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch6_eoc_prob10}
\alias{ch6_eoc_prob10}
\title{Chapter 6, Exercise 10}
\format{
A data frame with 3 rows and 2 variables.
 \describe{
  \item{B}{Stretch of fishing line under 3.5 kg load, Brand B}
  \item{D}{Stretch of fishing line under 3.5 kg load, Brand D}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 428
}
\usage{
data(ch6_eoc_prob10)
}
\description{
Chapter 6, Exercise 10
}
\examples{
data(ch6_eoc_prob10)
str(ch6_eoc_prob10)

}
\keyword{datasets}
