% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch6_eoc_prob11}
\alias{ch6_eoc_prob11}
\title{Chapter 6, Exercise 11}
\format{
A data frame with 10 rows and 1 variable.
 \describe{
  \item{y}{Purity of oxygen (in percent) delivered by an industrial air products supplier}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 430
}
\usage{
data(ch6_eoc_prob11)
}
\description{
Chapter 6, Exercise 11
}
\note{
Transformation of data is left to the reader.
}
\examples{
data(ch6_eoc_prob11)
str(ch6_eoc_prob11)

}
\keyword{datasets}
