% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch6_eoc_prob13}
\alias{ch6_eoc_prob13}
\title{Chapter 6, Exercise 13}
\format{
A data frame with 5 rows and 2 variables.
 \describe{
  \item{octane_87}{Miles per gallon with 87 octane fuel}
  \item{octane_90}{Miles per gallon with 90 octane fuel}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 431
}
\usage{
data(ch6_eoc_prob13)
}
\description{
Chapter 6, Exercise 13
}
\examples{
data(ch6_eoc_prob13)
str(ch6_eoc_prob13)

}
\keyword{datasets}
