% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch6_eoc_prob15}
\alias{ch6_eoc_prob15}
\title{Chapter 6, Exercise 15}
\format{
A data frame with 7 rows and 3 variables.
 \describe{
  \item{wire}{Wire ID}
  \item{C0.0}{Resistivity of copper wire at 0.0 degrees C}
  \item{C21.8}{Resistivity of copper wire at 21.8 degrees C}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 431
}
\usage{
data(ch6_eoc_prob15)
}
\description{
Chapter 6, Exercise 15
}
\examples{
data(ch6_eoc_prob15)
str(ch6_eoc_prob15)

}
\keyword{datasets}
