% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch6_eoc_prob16}
\alias{ch6_eoc_prob16}
\title{Chapter 6, Exercise 16}
\format{
A data frame with 7 rows and 1 variable.
 \describe{
  \item{resist}{Resistivity of aluminum wire at 21.8 degrees C}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 432
}
\usage{
data(ch6_eoc_prob16)
}
\description{
Chapter 6, Exercise 16
}
\examples{
data(ch6_eoc_prob16)
str(ch6_eoc_prob16)

}
\keyword{datasets}
