% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch6_eoc_prob20}
\alias{ch6_eoc_prob20}
\title{Chapter 6, Exercise 20}
\format{
A data frame with 20 rows and 3 variables.
 \describe{
  \item{Setting}{1 = improved, 2 = original}
  \item{Gear}{Gear ID within setting}
  \item{y1}{Flatness of gear (mm of distortion)}
  \item{y2}{Concentricity of gear (mm of distortion)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 433
}
\usage{
data(ch6_eoc_prob20)
}
\description{
Chapter 6, Exercise 20
}
\examples{
data(ch6_eoc_prob20)
str(ch6_eoc_prob20)

}
\keyword{datasets}
