% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch6_eoc_prob24}
\alias{ch6_eoc_prob24}
\title{Chapter 6, Exercise 24}
\format{
A data frame with 5 rows and 2 variables.
 \describe{
  \item{Rect}{Gripping pressure (manometers) delivered to rectangular objects}
  \item{Circ}{Gripping pressure (manometers) delivered to circular objects}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 435
}
\usage{
data(ch6_eoc_prob24)
}
\description{
Chapter 6, Exercise 24
}
\examples{
data(ch6_eoc_prob24)
str(ch6_eoc_prob24)

}
\keyword{datasets}
