% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch6_eoc_prob35}
\alias{ch6_eoc_prob35}
\title{Chapter 6, Exercise 35}
\format{
A data frame with 4 rows and 2 variables.
 \describe{
  \item{Van1}{Tilttable ratio for minivan 1}
  \item{Van2}{Tilttable ratio for minivan 2}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 439
}
\usage{
data(ch6_eoc_prob35)
}
\description{
Chapter 6, Exercise 35
}
\examples{
data(ch6_eoc_prob35)
str(ch6_eoc_prob35)

}
\keyword{datasets}
