% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch6_sec1_prob04}
\alias{ch6_sec1_prob04}
\title{Section 6.1, Exercise 4}
\format{
A data frame with 50 rows and 1 variable.
 \describe{
  \item{Diameter}{Diameter of No. 10 machine screws (mm)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 265
}
\usage{
data(ch6_sec1_prob04)
}
\description{
Section 6.1, Exercise 4
}
\note{
Data are shown as individual measurements and reused in Section 6.2, Exercise 4.
}
\examples{
data(ch6_sec1_prob04)
str(ch6_sec1_prob04)

}
\keyword{datasets}
