% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch6_xmp10}
\alias{ch6_xmp10}
\title{Chapter 6, Example 10}
\format{
A data frame with 24 rows and 2 variableS.
 \describe{
  \item{molded}{Packing weights for molded pieces (g)}
  \item{crushed}{Packing weights for crushed pieces (g)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 375
}
\usage{
data(ch6_xmp10)
}
\description{
Chapter 6, Example 10
}
\examples{
data(ch6_xmp10)
str(ch6_xmp10)

}
\keyword{datasets}
