% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch7_eoc_prob01}
\alias{ch7_eoc_prob01}
\title{Chapter 7, Exercise 1}
\format{
A data frame with 5 rows and 3 variables.
 \describe{
  \item{Lead_4H}{Grams of load on 4H pencil lead applied at failure}
  \item{Lead_H}{Grams of load on H pencil lead applied at failure}
  \item{Lead_B}{Grams of load on B pencil lead applied at failure}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 533
}
\usage{
data(ch7_eoc_prob01)
}
\description{
Chapter 7, Exercise 1
}
\examples{
data(ch7_eoc_prob01)
str(ch7_eoc_prob01)

}
\keyword{datasets}
