% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch7_eoc_prob02}
\alias{ch7_eoc_prob02}
\title{Chapter 7, Exercise 2}
\format{
A data frame with 20 rows and 3 variables.
 \describe{
  \item{Sample}{Sample ID}
  \item{x}{Mean of 4 diameter measurements on metal parts}
  \item{R}{Range of 4 diameter measurements on metal parts}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 534
}
\usage{
data(ch7_eoc_prob02)
}
\description{
Chapter 7, Exercise 2
}
\examples{
data(ch7_eoc_prob02)
str(ch7_eoc_prob02)

}
\keyword{datasets}
