% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch7_eoc_prob03}
\alias{ch7_eoc_prob03}
\title{Chapter 7, Exercise 3}
\format{
A data frame with 3 rows and 3 variables.
 \describe{
  \item{Sample}{Sample ID}
  \item{x}{Mean of 3 measurements water weight loss of fabric}
  \item{R}{Standard deviation of 3 measurements water weight loss of fabric}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 534
}
\usage{
data(ch7_eoc_prob03)
}
\description{
Chapter 7, Exercise 3
}
\examples{
data(ch7_eoc_prob03)
str(ch7_eoc_prob03)

}
\keyword{datasets}
