% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch7_eoc_prob04}
\alias{ch7_eoc_prob04}
\title{Chapter 7, Exercise 4}
\format{
A data frame with 8 rows and 3 variables.
 \describe{
  \item{i}{Construction adhesive ID}
  \item{ybar}{Mean strength of 8 tested joints}
  \item{s}{Standard deviation strength of 8 tested joints}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 535
}
\usage{
data(ch7_eoc_prob04)
}
\description{
Chapter 7, Exercise 4
}
\examples{
data(ch7_eoc_prob04)
str(ch7_eoc_prob04)

}
\keyword{datasets}
