% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch7_eoc_prob12}
\alias{ch7_eoc_prob12}
\title{Chapter 7, Exercise 12}
\format{
A data frame with 20 rows and 3 variables.
 \describe{
  \item{Widget}{Widget ID}
  \item{A}{Type A nonconformance}
  \item{B}{Type B nonconformance}
  \item{Total}{Total nonconformances}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 538
}
\usage{
data(ch7_eoc_prob12)
}
\description{
Chapter 7, Exercise 12
}
\examples{
data(ch7_eoc_prob12)
str(ch7_eoc_prob12)

}
\keyword{datasets}
