% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch7_eoc_prob15}
\alias{ch7_eoc_prob15}
\title{Chapter 7, Exercise 15}
\format{
A data frame with 35 rows and 2 variables.
 \describe{
  \item{Tank}{Space shuttle fuel tank ID}
  \item{Nonconf}{Number of  nonconformances}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 540
}
\usage{
data(ch7_eoc_prob15)
}
\description{
Chapter 7, Exercise 15
}
\examples{
data(ch7_eoc_prob15)
str(ch7_eoc_prob15)

}
\keyword{datasets}
