% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch7_eoc_prob16}
\alias{ch7_eoc_prob16}
\title{Chapter 7, Exercise 16}
\format{
A data frame with 25 rows and 4 variables.
 \describe{
  \item{SampleD1}{ID for sample taken on Day 1}
  \item{NonconD1}{Number of nonconforming in sample from Day 1}
  \item{SampleD2}{ID for sample taken on Day 2}
  \item{NonconD2}{Number of nonconforming in sample from Day 2}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 540
}
\usage{
data(ch7_eoc_prob16)
}
\description{
Chapter 7, Exercise 16
}
\examples{
data(ch7_eoc_prob16)
str(ch7_eoc_prob16)

}
\keyword{datasets}
