% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch7_eoc_prob17}
\alias{ch7_eoc_prob17}
\title{Chapter 7, Exercise 17}
\format{
A data frame with 15 rows and 2 variables.
 \describe{
  \item{Sample}{Sample ID}
  \item{Nonconf}{Number of nonconformances}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 541
}
\usage{
data(ch7_eoc_prob17)
}
\description{
Chapter 7, Exercise 17
}
\examples{
data(ch7_eoc_prob17)
str(ch7_eoc_prob17)

}
\keyword{datasets}
