% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch7_eoc_prob18}
\alias{ch7_eoc_prob18}
\title{Chapter 7, Exercise 18}
\format{
A data frame with 15 rows and 2 variables.
 \describe{
  \item{Temp}{Boiling point of 10W30 motor oil (F)}
  \item{Brand}{Brand of motor oil}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 541
}
\usage{
data(ch7_eoc_prob18)
}
\description{
Chapter 7, Exercise 18
}
\examples{
data(ch7_eoc_prob18)
str(ch7_eoc_prob18)

}
\keyword{datasets}
