% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch7_eoc_prob19}
\alias{ch7_eoc_prob19}
\title{Chapter 7, Exercise 19}
\format{
A data frame with 60 rows and 2 variables.
 \describe{
  \item{Sample}{Sample ID}
  \item{Lengths of metal rods above 33.69 in.}{}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 542
}
\usage{
data(ch7_eoc_prob19)
}
\description{
Chapter 7, Exercise 19
}
\examples{
data(ch7_eoc_prob19)
str(ch7_eoc_prob19)

}
\keyword{datasets}
