% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch7_eoc_prob21}
\alias{ch7_eoc_prob21}
\title{Chapter 7, Exercise 21}
\format{
A data frame with 12 rows and 3 variables.
 \describe{
  \item{Month}{Month of study}
  \item{Injuries}{Number of disabling injuries suffered}
  \item{Manhr}{Millions of man-hours worked}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 543
}
\usage{
data(ch7_eoc_prob21)
}
\description{
Chapter 7, Exercise 21
}
\examples{
data(ch7_eoc_prob21)
str(ch7_eoc_prob21)

}
\keyword{datasets}
