% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch7_eoc_prob22}
\alias{ch7_eoc_prob22}
\title{Chapter 7, Exercise 22}
\format{
A data frame with 15 rows and 2 variables.
 \describe{
  \item{Trials}{Force required to move cutter arm of a paper trimmer}
  \item{Type}{Type of paper}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 543
}
\usage{
data(ch7_eoc_prob22)
}
\description{
Chapter 7, Exercise 22
}
\examples{
data(ch7_eoc_prob22)
str(ch7_eoc_prob22)
}
\keyword{datasets}
