% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch7_eoc_prob24}
\alias{ch7_eoc_prob24}
\title{Chapter 7, Exercise 24}
\format{
A data frame with 12 rows and 2 variables.
 \describe{
  \item{Depth}{Depth of cut}
  \item{Pulse}{Coded number of pulses of laser beam}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 544
}
\usage{
data(ch7_eoc_prob24)
}
\description{
Chapter 7, Exercise 24
}
\examples{
data(ch7_eoc_prob24)
str(ch7_eoc_prob24)
}
\keyword{datasets}
