% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch7_sec1_prob2}
\alias{ch7_sec1_prob2}
\title{Section 7.1, Exercise 2}
\format{
A data frame with 5 rows and 4 variables.
 \describe{
  \item{Van1}{Tilttable ratio for Van 1}
  \item{Van2}{Tilttable ratio for Van 2}
  \item{Van3}{Tilttable ratio for Van 3}
  \item{Van4}{Tilttable ratio for Van 4}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 460
}
\usage{
data(ch7_sec1_prob2)
}
\description{
Section 7.1, Exercise 2
}
\note{
Van #3 was tested five times while the others were tested four times each. No data are missing.
}
\examples{
data(ch7_sec1_prob2)
str(ch7_sec1_prob2)

}
\keyword{datasets}
