% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch7_sec5_prob2}
\alias{ch7_sec5_prob2}
\title{Section 7.5, Exercise 2}
\format{
A data frame with 95 rows and 2 variables.
 \describe{
  \item{Sample}{Sample ID}
  \item{Length}{Thread length of U-bolt}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 516
}
\usage{
data(ch7_sec5_prob2)
}
\description{
Section 7.5, Exercise 2
}
\examples{
data(ch7_sec5_prob2)
str(ch7_sec5_prob2)

}
\keyword{datasets}
