% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch7_sec6_prob1}
\alias{ch7_sec6_prob1}
\title{Section 7.6, Exercise 1}
\format{
A data frame with 20 rows and 4 variables.
 \describe{
  \item{Sample}{Sample ID}
  \item{Defects}{Number of defectives}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 532
}
\usage{
data(ch7_sec6_prob1)
}
\description{
Section 7.6, Exercise 1
}
\examples{
data(ch7_sec5_prob6)
str(ch7_sec5_prob6)

}
\keyword{datasets}
