% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch7_sec6_prob2}
\alias{ch7_sec6_prob2}
\title{Section 7.6, Exercise 2}
\format{
A data frame with 18 rows and 3 variables.
 \describe{
  \item{Date}{Sample ID}
  \item{Tested}{Number of radiators tested}
  \item{Leaks}{Number of leaks}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 532
}
\usage{
data(ch7_sec6_prob2)
}
\description{
Section 7.6, Exercise 2
}
\examples{
data(ch7_sec6_prob2)
str(ch7_sec6_prob2)

}
\keyword{datasets}
