% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch7_sec6_prob3}
\alias{ch7_sec6_prob3}
\title{Section 7.6, Exercise 3}
\format{
A data frame with 8 rows and 3 variables.
 \describe{
  \item{i}{Sample ID}
  \item{k}{Number of units inspected}
  \item{u}{Defects per unit}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 532
}
\usage{
data(ch7_sec6_prob3)
}
\description{
Section 7.6, Exercise 3
}
\examples{
data(ch7_sec6_prob3)
str(ch7_sec6_prob3)

}
\keyword{datasets}
