% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch7_sec6_prob4}
\alias{ch7_sec6_prob4}
\title{Section 7.6, Exercise 4}
\format{
A data frame with 10 rows and 3 variables.
 \describe{
  \item{Sample}{Sample ID}
  \item{Size}{Sample size}
  \item{Nonconf}{Number nonconforming}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 533
}
\usage{
data(ch7_sec6_prob4)
}
\description{
Section 7.6, Exercise 4
}
\examples{
data(ch7_sec6_prob4)
str(ch7_sec6_prob4)

}
\keyword{datasets}
