% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch8_eoc_prob04}
\alias{ch8_eoc_prob04}
\title{Chapter 8, Exercise 4}
\format{
A data frame with 12 rows and 4 variables.
 \describe{
  \item{ybar}{Mean viscosity of motor oil}
  \item{s}{Standard deviation of viscosity of motor oil}
  \item{brand}{Brand of motor oil}
  \item{oil}{Type of motor oil (1 = 10W30, 2 = SAE 30, 3 = 10W40, 4 = 20W50)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 633
}
\usage{
data(ch8_eoc_prob04)
}
\description{
Chapter 8, Exercise 4
}
\examples{
data(ch8_eoc_prob04)
str(ch8_eoc_prob04)
}
\keyword{datasets}
