% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch8_eoc_prob06}
\alias{ch8_eoc_prob06}
\title{Chapter 8, Exercise 6}
\format{
A data frame with 8 rows and 4 variables.
 \describe{
  \item{ybar}{Mean of compressive strength for 5 specimens of Type III cement concrete (MPa)}
  \item{s}{Standard deviation of compressive strength}
  \item{Ratio}{Water-cement ration}
  \item{Temp}{Curing temperature (C)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 634
}
\usage{
data(ch8_eoc_prob06)
}
\description{
Chapter 8, Exercise 6
}
\examples{
data(ch8_eoc_prob06)
str(ch8_eoc_prob06)
}
\keyword{datasets}
