% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch8_eoc_prob08}
\alias{ch8_eoc_prob08}
\title{Chapter 8, Exercise 8}
\format{
A data frame with 8 rows and 5 variables.
 \describe{
  \item{A}{Crankshaft material: cast modular iron (-1), forged steel (+1)}
  \item{B}{Bearing material: aluminum (-1), copper-lead (+1)}
  \item{C}{Debris added to oil: none (-1), 5.5 g SAE fine dust (+1)}
  \item{y1}{Rod journal wear (micrometers)}
  \item{y2}{Main journal wear (micrometers)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 635
}
\usage{
data(ch8_eoc_prob08)
}
\description{
Chapter 8, Exercise 8
}
\examples{
data(ch8_eoc_prob08)
str(ch8_eoc_prob08)
}
\keyword{datasets}
