% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch8_eoc_prob11}
\alias{ch8_eoc_prob11}
\title{Chapter 8, Exercise 11}
\format{
A data frame with 16 rows and 6 variables.
 \describe{
  \item{A}{Lead type: standard (-1), new (+1)}
  \item{B}{Plant: 1 (-1), 2 (+1)}
  \item{C}{Machine type: standard (-1), high speed (+1)}
  \item{D}{Shift: first (-1), second (+1)}
  \item{ybar}{Mean number of leads missed per hour}
  \item{ssquare}{Variance of leads missed per hour}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 637
}
\usage{
data(ch8_eoc_prob11)
}
\description{
Chapter 8, Exercise 11
}
\examples{
data(ch8_eoc_prob11)
str(ch8_eoc_prob11)
}
\keyword{datasets}
