% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch8_eoc_prob12}
\alias{ch8_eoc_prob12}
\title{Chapter 8, Exercise 12}
\format{
A data frame with 27 rows and 4 variables.
 \describe{
  \item{Type}{Can type: cola (1), beer (2), soup (3)}
  \item{Temp}{0C (1), 22C (2), 200C (3)}
  \item{trt}{Can type-temperature combination}
  \item{y}{Force required to buckle can (lb)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 638
}
\usage{
data(ch8_eoc_prob12)
}
\description{
Chapter 8, Exercise 12
}
\examples{
data(ch8_eoc_prob12)
str(ch8_eoc_prob12)
}
\keyword{datasets}
