% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch8_eoc_prob17}
\alias{ch8_eoc_prob17}
\title{Chapter 8, Exercise 17}
\format{
A data frame with 8 rows and 1 variable.
 \describe{
  \item{means}{Artificial means}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 639
}
\usage{
data(ch8_eoc_prob17)
}
\description{
Chapter 8, Exercise 17
}
\examples{
data(ch8_eoc_prob17)
str(ch8_eoc_prob17)
}
\keyword{datasets}
