% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch8_eoc_prob19}
\alias{ch8_eoc_prob19}
\title{Chapter 8, Exercise 19}
\format{
A data frame with 16 rows and 6 variables.
 \describe{
  \item{A}{Conveyor speed: -1 or +1}
  \item{B}{Preheat temperature: -1 or +1}
  \item{C}{Solder temperature: -1 or +1}
  \item{D}{Conveyor angle: -1 or +1}
  \item{E}{Flux concentration: -1 or +1}
  \item{y}{Number of faults per 100 solder joints on circuit boards}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 640
}
\usage{
data(ch8_eoc_prob19)
}
\description{
Chapter 8, Exercise 19
}
\note{
Actual levels of factors were not given.
}
\examples{
data(ch8_eoc_prob19)
str(ch8_eoc_prob19)
}
\keyword{datasets}
