% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch8_eoc_prob20}
\alias{ch8_eoc_prob20}
\title{Chapter 8, Exercise 20}
\format{
A data frame with 16 rows and 10 variables.
 \describe{
  \item{A}{Arsenic flow rate (percent): 55 (-1) or 59 (+1)}
  \item{B}{Deposition temperature (C): 1210 (-1) or 1220 (+1)}
  \item{C}{Code of wafers: 668G4 (-1) or 678G4 (+1)}
  \item{D}{Susceptor rotation: continuous (-1) or oscillating (+1)}
  \item{E}{Deposition time: high (-1) or low (+1)}
  \item{F}{HC1 Etch temperature (C): 1180 (-1) or 1215 (+1)}
  \item{G}{HC1 Flow rate (percent): 10 (-1) or 14 (+1)}
  \item{H}{Nozzle position: 2 (-1) or 6 (+1)}
  \item{y1}{Mean of thickness measurements (mm)}
  \item{y2}{Logarithm of variance of thickness measurements}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 641
}
\usage{
data(ch8_eoc_prob20)
}
\description{
Chapter 8, Exercise 20
}
\examples{
data(ch8_eoc_prob20)
str(ch8_eoc_prob20)
}
\keyword{datasets}
