% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch8_eoc_prob21}
\alias{ch8_eoc_prob21}
\title{Chapter 8, Exercise 21}
\format{
A data frame with 16 rows and 7 variables.
 \describe{
  \item{A}{Bulk density: low (-1) or high (+1)}
  \item{B}{Moisture: low (-1) or high (+1)}
  \item{C}{Crammer current: low (-1) or high (+1)}
  \item{D}{Extruder screw speed: low (-1) or high (+1)}
  \item{E}{Front-end temperature: low (-1) or high (+1)}
  \item{F}{Back-end temperature: low (-1) or high (+1)}
  \item{ybar}{Mean output}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 642
}
\usage{
data(ch8_eoc_prob21)
}
\description{
Chapter 8, Exercise 21
}
\examples{
data(ch8_eoc_prob21)
str(ch8_eoc_prob21)
}
\keyword{datasets}
