% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch8_eoc_prob22}
\alias{ch8_eoc_prob22}
\title{Chapter 8, Exercise 22}
\format{
A data frame with 32 rows and 10 variables.
 \describe{
  \item{A}{Preheat temperature: 180 (-1) or 220 (+1)}
  \item{B}{Solder wave height: 250 (-1) or 400 (+1)}
  \item{C}{Wave temperature: 490 (-1) or 510 (+1)}
  \item{D}{Conveyor angle: 5.0 (-1) or 6.1 (+1)}
  \item{E}{Flux type: A857 (-1) or K192 (+1)}
  \item{F}{Direction of boards: 0 (-1) or 90 (+1)}
  \item{G}{Wave width: 2.25 (-1) or 3.00 (+1)}
  \item{H}{Conveyor speed: 3.5 (-1) or 6.0 (+1)}
  \item{y1}{Lead shorts per circuit board}
  \item{y2}{Touch shorts per circuit board}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 643
}
\usage{
data(ch8_eoc_prob22)
}
\description{
Chapter 8, Exercise 22
}
\examples{
data(ch8_eoc_prob22)
str(ch8_eoc_prob22)
}
\keyword{datasets}
