% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch8_eoc_prob24}
\alias{ch8_eoc_prob24}
\title{Chapter 8, Exercise 24}
\format{
A data frame with 12 rows and 5 variables.
 \describe{
  \item{Day}{Day of experimental run}
  \item{Flow}{Total feed flow: 50 (-1) or 180 (+1)}
  \item{Temp}{Reactor wall temperature (F): 470 (-1) or 520 (+1)}
  \item{Ratio}{Feed ratio: 4 (-1) or 8 (+1)}
  \item{y}{Percent conversion of butane}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 645
}
\usage{
data(ch8_eoc_prob24)
}
\description{
Chapter 8, Exercise 24
}
\examples{
data(ch8_eoc_prob24)
str(ch8_eoc_prob24)
}
\keyword{datasets}
