% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch8_eoc_prob25}
\alias{ch8_eoc_prob25}
\title{Chapter 8, Exercise 25}
\format{
A data frame with 8 rows and 4 variables.
 \describe{
  \item{A}{Speed (rpm): 2500 (-1) or 4500 (+1)}
  \item{B}{Feed (in/rev): 0.003 (-1) or 0.009 (+1)}
  \item{C}{Tool condition: new (-1) or used (+1)}
  \item{ybar}{Mean surface finish measurement (m = 2)}
  \item{s}{Standard deviation surface finish measurement}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 646
}
\usage{
data(ch8_eoc_prob25)
}
\description{
Chapter 8, Exercise 25
}
\examples{
data(ch8_eoc_prob25)
str(ch8_eoc_prob25)
}
\keyword{datasets}
