% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch8_eoc_prob26}
\alias{ch8_eoc_prob26}
\title{Chapter 8, Exercise 26}
\format{
A data frame with 16 rows and 6 variables.
 \describe{
  \item{A}{Chamber pressure: 8 (-1) or 9 (+1)}
  \item{B}{H2 flow: 500 (-1) or 1000 (+1)}
  \item{C}{SiH4 flow: 15 (-1) or 25 (+1)}
  \item{D}{WF6 flow: 50 (-1) or 60 (+1)}
  \item{y1}{Percent step coverage}
  \item{y2}{Average sheet resistance}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 647
}
\usage{
data(ch8_eoc_prob26)
}
\description{
Chapter 8, Exercise 26
}
\examples{
data(ch8_eoc_prob26)
str(ch8_eoc_prob26)
}
\keyword{datasets}
