% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch8_eoc_prob28}
\alias{ch8_eoc_prob28}
\title{Chapter 8, Exercise 28}
\format{
A data frame with 15 rows and 1 variable.
 \describe{
  \item{Fit}{Separated clear volume for a suspension of antibiotic (percent)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 648
}
\usage{
data(ch8_eoc_prob28)
}
\description{
Chapter 8, Exercise 28
}
\examples{
data(ch8_eoc_prob28)
str(ch8_eoc_prob28)
}
\keyword{datasets}
