% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch8_eoc_prob29}
\alias{ch8_eoc_prob29}
\title{Chapter 8, Exercise 29}
\format{
A data frame with 8 rows and 4 variables.
 \describe{
  \item{A}{Tube height: low (-1) or high (+1)}
  \item{B}{Tube width: low (-1) or high (+1)}
  \item{C}{Paint viscosity: low (-1) or high (+1)}
  \item{y}{Paint coating thickness (mm)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 649
}
\usage{
data(ch8_eoc_prob29)
}
\description{
Chapter 8, Exercise 29
}
\examples{
data(ch8_eoc_prob29)
str(ch8_eoc_prob29)
}
\keyword{datasets}
