% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch8_sec1_prob1}
\alias{ch8_sec1_prob1}
\title{Section 8.1, Exercise 1}
\format{
A data frame with 18 rows and 4 variables.
 \describe{
  \item{Strg}{Shear strength of joint (lb)}
  \item{wood}{1 = pine, 2 = fir}
  \item{glue}{1 = white, 2 = carpenters, 3 = cascamite}
  \item{trt}{wood-glue combination}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 567
}
\usage{
data(ch8_sec1_prob1)
}
\description{
Section 8.1, Exercise 1
}
\examples{
data(ch8_sec1_prob1)
str(ch8_sec1_prob1)
}
\keyword{datasets}
