% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch8_sec1_prob2}
\alias{ch8_sec1_prob2}
\title{Section 8.1, Exercise 2}
\format{

}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 568
}
\usage{
data(ch8_sec1_prob2)
}
\description{
Section 8.1, Exercise 2
}
\examples{
data(ch8_sec1_prob2)
str(ch8_sec1_prob2)
}
\keyword{datasets}
