% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch8_sec2_prob3}
\alias{ch8_sec2_prob3}
\title{Section 8.2, Exercise 3}
\format{
A data frame with 16 rows and 5 variables.
 \describe{
  \item{A}{Fabric tested: sateen (-1), monk's cloth (+1)}
  \item{B}{Treatment: X (-1), Y (+1)}
  \item{C}{Laundering condition: before (-1), after (+1)}
  \item{D}{Direction of test: warp (-1), fill (+1)}
  \item{y}{Inches burned on standard-size sample in flame test}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 590
}
\usage{
data(ch8_sec2_prob3)
}
\description{
Section 8.2, Exercise 3
}
\examples{
data(ch8_sec2_prob3)
str(ch8_sec2_prob3)
}
\keyword{datasets}
