% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch8_sec3_prob2}
\alias{ch8_sec3_prob2}
\title{Section 8.3, Exercise 2}
\format{
A data frame with 16 rows and 6 variables.
 \describe{
  \item{A}{Constant velocity: 0.6 in/sec (-1), 1.2 in/sec (+1)}
  \item{B}{Temperature: 150C (-1), 300C (+1)}
  \item{C}{Bond force: 80 g (-1), 120 g (+1)}
  \item{D}{Ultrasonic power: 120 mW (-1), 200 mW (+1)}
  \item{E}{Bond time: 10 ms (-1), 20 ms (+1)}
  \item{y}{Force required to pull wire bonds (g)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 611
}
\usage{
data(ch8_sec3_prob2)
}
\description{
Section 8.3, Exercise 2
}
\examples{
data(ch8_sec3_prob2)
str(ch8_sec3_prob2)
}
\keyword{datasets}
