% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch9_eoc_prob03}
\alias{ch9_eoc_prob03}
\title{Chapter 9, Exercise 3}
\format{
A data frame with 10 rows and 5 variables.
 \describe{
  \item{x1}{Force (g)}
  \item{x2}{Power (mw)}
  \item{x3}{Temp (C)}
  \item{x4}{Time (ms)}
  \item{y}{Ball bond shear strength}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 722
}
\usage{
data(ch9_eoc_prob03)
}
\description{
Chapter 9, Exercise 3
}
\note{
Data also used in Chapter 9, Exercise 4d
}
\examples{
data(ch9_eoc_prob03)
str(ch9_eoc_prob03)
}
\keyword{datasets}
