% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch9_eoc_prob07}
\alias{ch9_eoc_prob07}
\title{Chapter 9, Exercise 7}
\format{
A data frame with 80 rows and 2 variables.
 \describe{
  \item{y}{Weights of portions of 14-gram samples of oat seeds}
  \item{Spacing}{Door spacing of cleaning chamber of thresher}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 725
}
\usage{
data(ch9_eoc_prob07)
}
\description{
Chapter 9, Exercise 7
}
\examples{
data(ch9_eoc_prob05)
str(ch9_eoc_prob05)
}
\keyword{datasets}
