% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch9_sec3_prob1}
\alias{ch9_sec3_prob1}
\title{Section 9.3, Exercise 1}
\format{
A data frame with 48 rows and 4 variables.
 \describe{
  \item{x1}{Time (min) of tempering stainless steel slugs}
  \item{x2}{Temperature (F) of tempering stainless steel slugs}
  \item{y}{Increse in hardness}
  \item{trt}{Time-temperature combination}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 719
}
\usage{
data(ch9_sec3_prob1)
}
\description{
Section 9.3, Exercise 1
}
\examples{
data(ch9_sec3_prob1)
str(ch9_sec3_prob1)
}
\keyword{datasets}
