% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table1_01}
\alias{table1_01}
\title{Table 1.1}
\format{
A data frame with 39 rows and 2 variables.
 \describe{
  \item{Laid}{Thrust face runouts for laid gears (0.0001 in.)}
  \item{Hung}{Thrust face runouts for hung gears (0.0001 in.)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 3
}
\usage{
data(table1_01)
}
\description{
Table 1.1
}
\note{
Data used in Section 2.2, Exercise 1; Section 3.4, Exercise 2; Chapter 6, Exercise 2
}
\examples{
data(table1_01)
str(table1_01)

# reshape data for analysis
gears <- table1_01 |>
  tidyr::pivot_longer(cols = everything(),
                      values_to = "Runout",
                      names_to = "Treatment") |>
  dplyr::mutate(Treatment = factor(Treatment,
                                   levels = c("Laid", "Hung"))) |>
  # one laid gear was dropped from study; remove NA value
  dplyr::filter(!is.na(Runout))

# dot plot of runouts by hung or laid
mosaic::dotPlot(~ Runout | Treatment, data = gears)

# mean runout by treatment, mosaic version
mosaic::mean_(Runout ~ Treatment, data = gears)

# mean runout by treatment, tidyverse version
gears |>
  dplyr::summarize(runout_mean = mean(Runout), .by = Treatment)

# summary statistics of runout by treatment
mosaic::favstats(Runout ~ Treatment, data = gears)
}
\keyword{datasets}
