% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table1_04}
\alias{table1_04}
\title{Table 1.4}
\format{
A data frame with 23 rows and 2 variables.
 \describe{
  \item{point}{Time interval (1/60 sec.)}
  \item{displacement}{Displacement from initial position (mm)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) Basic Engineering Data Collection and Analysis
\url{https://www.iastatedigitalpress.com/plugins/books/127/}
}
\usage{
data(table1_04)
}
\description{
Table 1.4
}
\examples{
data(table1_04)
str(table1_04)

}
\keyword{datasets}
