% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table2_01}
\alias{table2_01}
\title{Table 2.1}
\format{
A data frame with 18 rows and 4 variables
 \describe{
  \item{sample}{Sample taken}
  \item{item}{Item within sample}
  \item{mass}{Mass of bottle (g)}
  \item{width}{Width of bottle (g)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) Basic Engineering Data Collection and Analysis
\url{https://www.iastatedigitalpress.com/plugins/books/127/}
}
\usage{
data(table2_01)
}
\description{
Table 2.1
}
\examples{
data(table2_01)
str(table2_01)

}
\keyword{datasets}
