% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table3_01}
\alias{table3_01}
\title{Table 3.1}
\format{
A data frame with 20 rows and 2 variables.
 \describe{
  \item{G230}{Bullet penetration depths for 230 grain jacketed bullets (mm)}
  \item{G200}{Bullet penetration depths for 200 grain jacketed bullets (mm)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 67
}
\usage{
data(table3_01)
}
\description{
Table 3.1
}
\note{
Data also used in Section 3.1, Exercise 2
}
\examples{
data(table3_01)
str(table3_01)

}
\keyword{datasets}
