% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table3_04}
\alias{table3_04}
\title{Table 3.4}
\format{
A data frame with 34 rows and 3 variables.
 \describe{
  \item{component}{Component ID}
  \item{bolt3_torque}{Torque to loosen bolt 3 (ft-lb)}
  \item{bolt4_torque}{Torque to loosen bolt 4 (ft-lb)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 74
}
\usage{
data(table3_04)
}
\description{
Table 3.4
}
\examples{
data(table3_04)
str(table3_04)

}
\keyword{datasets}
