% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table3_05}
\alias{table3_05}
\title{Table 3.5}
\format{
A data frame with 30 rows and 2 variables.
 \describe{
  \item{joint}{Joint ID}
  \item{diameter}{(inches above nominal)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) Basic Engineering Data Collection and Analysis
\url{https://www.iastatedigitalpress.com/plugins/books/127/}
}
\usage{
data(table3_05)
}
\description{
Table 3.5
}
\examples{
data(table3_05)
str(table3_05)

}
\keyword{datasets}
