% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table3_06}
\alias{table3_06}
\title{Table 3.6}
\format{
A data frame with 10 rows and 2 variables.
 \describe{
  \item{test}{Test ID}
  \item{strength}{(Breaking strength (g))}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) Basic Engineering Data Collection and Analysis
\url{https://www.iastatedigitalpress.com/plugins/books/127/}
}
\usage{
data(table3_06)
}
\description{
Table 3.6
}
\examples{
data(table3_06)
str(table3_06)

}
\keyword{datasets}
