% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table3_13}
\alias{table3_13}
\title{Table 3.13}
\format{
A data frame with 14 rows and 2 variables.
 \describe{
  \item{supplier}{Supplier ID}
  \item{pct_waste}{Percent waste by weight}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 94
}
\usage{
data(table3_13)
}
\description{
Table 3.13
}
\note{
Data used in Chapter 3, Section 2, Exercise 2
}
\examples{
data(table3_13)
str(table3_13)

}
\keyword{datasets}
