% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table3_14}
\alias{table3_14}
\title{Table 3.14}
\format{
A data frame with 25 rows and 3 variables.
 \describe{
  \item{sample}{Sample ID}
  \item{xbar}{Sample mean}
  \item{r}{Sample range}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}
}
\usage{
data(table3_14)
}
\description{
Table 3.14
}
\examples{
data(table3_14)
str(table3_14)

}
\keyword{datasets}
