% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table3_15}
\alias{table3_15}
\title{Table 3.15}
\format{
A data frame with 9 rows and 3 variables.
 \describe{
  \item{wood}{Wood type}
  \item{glue}{Glue type}
  \item{mean_strength}{Mean joint shear strength (lb)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}
}
\usage{
data(table3_15)
}
\description{
Table 3.15
}
\examples{
data(table3_15)
str(table3_15)

}
\keyword{datasets}
