% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table3_16}
\alias{table3_16}
\title{Table 3.16}
\format{
A data frame with 4 rows and 2 variables.
 \describe{
  \item{category}{Defect category}
  \item{connectors}{Number of sampled connectors}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) Basic Engineering Data Collection and Analysis
\url{https://www.iastatedigitalpress.com/plugins/books/127/}
}
\usage{
data(table3_16)
}
\description{
Table 3.16
}
\examples{
data(table3_16)
str(table3_16)

}
\keyword{datasets}
