% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table3_17}
\alias{table3_17}
\title{Table 3.17}
\format{
A data frame with 11 rows and 2 variables.
 \describe{
  \item{problem}{Description of problem}
  \item{glue}{Number of tools out of 100 with problem}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) Basic Engineering Data Collection and Analysis
\url{https://www.iastatedigitalpress.com/plugins/books/127/}
}
\usage{
data(table3_17)
}
\description{
Table 3.17
}
\examples{
data(table3_17)
str(table3_17)

}
\keyword{datasets}
