% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table3_18}
\alias{table3_18}
\title{Table 3.18}
\format{
A data frame with 11 rows and 2 variables.
 \describe{
  \item{job}{Job number}
  \item{type}{Type of action}
  \item{pct}{Percent requiring action}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) Basic Engineering Data Collection and Analysis
\url{https://www.iastatedigitalpress.com/plugins/books/127/}
}
\usage{
data(table3_18)
}
\description{
Table 3.18
}
\examples{
data(table3_18)
str(table3_18)

}
\keyword{datasets}
