% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table3_19}
\alias{table3_19}
\title{Table 3.19}
\format{
A data frame with 26 rows and 2 variables.
 \describe{
  \item{day}{Day number}
  \item{defects_truck}{Number of defects per truck}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 110
}
\usage{
data(table3_19)
}
\description{
Table 3.19
}
\examples{
data(table3_19)
str(table3_19)

}
\keyword{datasets}
