% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table3_20}
\alias{table3_20}
\title{Table 3.20}
\format{
A data frame with 4 rows and 3 variables.
 \describe{
  \item{shot_size}{Shot size}
  \item{mixture}{Percent of reground material}
  \item{n_conform}{Number conforming}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) Basic Engineering Data Collection and Analysis
\url{https://www.iastatedigitalpress.com/plugins/books/127/}
}
\usage{
data(table3_20)
}
\description{
Table 3.20
}
\examples{
data(table3_20)
str(table3_20)

}
\keyword{datasets}
