% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table4_01}
\alias{table4_01}
\title{Table 4.1}
\format{
A data frame with 15 rows and 2 variables.
 \describe{
  \item{pressure}{Pressing presssure (psi)}
  \item{density}{Density of cylinder (g/cc)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) Basic Engineering Data Collection and Analysis
\url{https://www.iastatedigitalpress.com/plugins/books/127/}
}
\usage{
data(table4_01)
}
\description{
Table 4.1
}
\examples{
data(table4_01)
str(table4_01)

}
\keyword{datasets}
