% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table4_03}
\alias{table4_03}
\title{Table 4.3}
\format{
A data frame with 18 rows and 2 variables.
 \describe{
  \item{percent}{Percent of ammonium phosphate}
  \item{strength}{Compressive strength (psi)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 133
}
\usage{
data(table4_03)
}
\description{
Table 4.3
}
\examples{
data(table4_03)
str(table4_03)

}
\keyword{datasets}
