% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table4_06}
\alias{table4_06}
\title{Table 4.6}
\format{
A data frame with 6 rows and 2 variables.
 \describe{
  \item{x}{Hypothetical predictor}
  \item{y}{Hypothetical response}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 139
}
\usage{
data(table4_06)
}
\description{
Table 4.6
}
\examples{
data(table4_06)
str(table4_06)

}
\keyword{datasets}
