% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table4_08}
\alias{table4_08}
\title{Table 4.8}
\format{
A data frame with 17 rows and 4 variables.
 \describe{
  \item{air}{Air flow}
  \item{water}{Cooling water inlet temperature}
  \item{acid}{Acid concentration}
  \item{stack}{Stack loss}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) Basic Engineering Data Collection and Analysis
\url{https://www.iastatedigitalpress.com/plugins/books/127/}
}
\usage{
data(table4_08)
}
\description{
Table 4.8
}
\note{
Brownlee's stack loss data, presented here with 17 cases instead of the usual 21. Complete data are available as `datasets::stackloss`.
}
\examples{
data(table4_08)
str(table4_08)

}
\references{
Brownlee, K.A. (1965) \emph{Statistical Theory and Methodology In Science and engineering} (second edition), New York: Wiley.
}
\keyword{datasets}
