% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table4_10}
\alias{table4_10}
\title{Table 4.10}
\format{
A data frame with 9 rows and 3 variables.
 \describe{
  \item{canard}{Canard position}
  \item{tail}{Tail position}
  \item{lift_drag}{Lift/drag ratio}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 156
}
\usage{
data(table4_10)
}
\description{
Table 4.10
}
\examples{
data(table4_10)
str(table4_10)

}
\keyword{datasets}
