% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table4_11}
\alias{table4_11}
\title{Table 4.11}
\format{
A data frame with 16 rows and 4 variables.
 \describe{
  \item{specimen}{Specimen ID}
  \item{joint}{Joint type}
  \item{wood}{Wood type}
  \item{strength}{Stress at failure (psi)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 164
}
\usage{
data(table4_11)
}
\description{
Table 4.11
}
\examples{
data(table4_11)
str(table4_11)

}
\keyword{datasets}
