% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table4_15}
\alias{table4_15}
\title{Table 4.15}
\format{
A data frame with 40 rows and 3 variables.
 \describe{
  \item{distance}{Flight distance of golf ball}
  \item{compression}{Compression of golf ball}
  \item{evening}{Evening of data collecton}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 173
}
\usage{
data(table4_15)
}
\description{
Table 4.15
}
\examples{
data(table4_15)
str(table4_15)

}
\keyword{datasets}
