% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table4_20}
\alias{table4_20}
\title{Table 4.20}
\format{
A data frame with 8 rows and 4 variables.
 \describe{
  \item{temp}{Autoclave temperature, 300F or 330F}
  \item{time}{Autoclave time, 4 hr or 12 hr}
  \item{span}{Time between product formation and autoclaving, 4 hr or 12 hr}
  \item{strength_mean}{Mean strength (psi)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 180
}
\usage{
data(table4_20)
}
\description{
Table 4.20
}
\examples{
data(table4_20)
str(table4_20)

}
\keyword{datasets}
