% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table4_24}
\alias{table4_24}
\title{Table 4.24}
\format{
A data frame with 16 rows and 2 variables.
 \describe{
  \item{combn}{Combination of factor levels}
  \item{y}{Rate of advance of drill}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 195
}
\usage{
data(table4_24)
}
\description{
Table 4.24
}
\note{
Units of y not given
}
\examples{
data(table4_24)
str(table4_24)

}
\keyword{datasets}
