% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table4_25}
\alias{table4_25}
\title{Table 4.25}
\format{
A data frame with 9 rows and 4 variables.
 \describe{
  \item{temp}{Condensation temperature (C)}
  \item{amt_B}{Amount of B (cc)}
  \item{yield}{Yield (g)}
  \item{fil_time}{Filtration time (sec)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 199
}
\usage{
data(table4_25)
}
\description{
Table 4.25
}
\examples{
data(table4_25)
str(table4_25)

}
\keyword{datasets}
