% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table5_01}
\alias{table5_01}
\title{Table 5.1}
\format{
A data frame with 10 rows and 2 variables.
 \describe{
  \item{z}{Torque (ft-lb) required to loosen bolt 3 on metal plate of a piece of heavy machinery}
  \item{n}{Frequency}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 224
}
\usage{
data(table5_01)
}
\description{
Table 5.1
}
\examples{
data(table5_01)
str(table5_01)

}
\keyword{datasets}
