% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table5_06}
\alias{table5_06}
\title{Table 5.6}
\format{
A data frame with 100 rows and 1 variable.
 \describe{
  \item{weight}{Weight of U.S. nickels (g)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 265
}
\usage{
data(table5_06)
}
\description{
Table 5.6
}
\note{
Data are shown as individual measurements in random order.
}
\examples{
data(table5_06)
str(table5_06)

}
\keyword{datasets}
