% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table5_08}
\alias{table5_08}
\title{Table 5.8}
\format{
A data frame with 25 rows and 1 variable.
 \describe{
  \item{thread_lgth}{Thread length of U-bolts (0.001 in over nominal)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 267
}
\usage{
data(table5_08)
}
\description{
Table 5.8
}
\note{
Data are shown as individual measurements in random order.
}
\examples{
data(table5_08)
str(table5_08)

}
\keyword{datasets}
