% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table5_09}
\alias{table5_09}
\title{Table 5.9}
\format{
A data frame with 20 rows and 1 variable.
 \describe{
  \item{voltage}{Failure voltage (kv/mm) for an electrical cable insulation subjected to increasing voltage stress}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 276
}
\usage{
data(table5_09)
}
\description{
Table 5.9
}
\examples{
data(table5_09)
str(table5_09)

}
\keyword{datasets}
