% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table6_04}
\alias{table6_04}
\title{Table 6.4}
\format{
A data frame with 10 rows and 1 variable.
 \describe{
  \item{lifetime}{Number of cycles to failure of springs under stress of 950 N/mm2}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 366
}
\usage{
data(table6_04)
}
\description{
Table 6.4
}
\examples{
data(table6_04)
str(table6_04)

}
\keyword{datasets}
