% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table6_05}
\alias{table6_05}
\title{Table 6.5}
\format{
A data frame with 5 rows and 3 variables.
 \describe{
  \item{piece}{Wood product piece ID}
  \item{leading}{Leading edge (in.)}
  \item{trailing}{Trailing edge (in.)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 370
}
\usage{
data(table6_05)
}
\description{
Table 6.5
}
\examples{
data(table6_05)
str(table6_05)

}
\keyword{datasets}
