% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table6_07}
\alias{table6_07}
\title{Table 6.7}
\format{
A data frame with 10 rows and 2 variables.
 \describe{
  \item{stress_950}{Number of cycles to failure of springs under stress of 950 N/mm2}
  \item{stress_900}{Number of cycles to failure of springs under stress of 900 N/mm2}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 379
}
\usage{
data(table6_07)
}
\description{
Table 6.7
}
\note{
Data reused in Section 6.6, Exercise 2
}
\examples{
data(table6_07)
str(table6_07)

}
\keyword{datasets}
