% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table6_08}
\alias{table6_08}
\title{Table 6.8}
\format{
A data frame with 20 rows and 1 variable.
 \describe{
  \item{dimension}{measurements of a dimension on 20 parts machined on a CNC lathe (0.0001 in. over nominal)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 389
}
\usage{
data(table6_08)
}
\description{
Table 6.8
}
\note{
Data shown as individual measurements in random order.
}
\examples{
data(table6_08)
str(table6_08)

}
\keyword{datasets}
