% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table6_09}
\alias{table6_09}
\title{Table 6.9}
\format{
A data frame with 15 rows and 2 variable.
 \describe{
  \item{hardness}{Rockwell hardness of steel specimens}
  \item{treatment}{Treatment of specimen: heat treated or cold rolled}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 395
}
\usage{
data(table6_09)
}
\description{
Table 6.9
}
\examples{
data(table6_09)
str(table6_09)

}
\keyword{datasets}
