% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table6_10}
\alias{table6_10}
\title{Table 6.10}
\format{
A data frame with 100 rows and 1 variable.
 \describe{
  \item{wgt}{Weight of newly minted U.S. pennies (g)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 417
}
\usage{
data(table6_10)
}
\description{
Table 6.10
}
\note{
Data shown as individual measurements in random order.
}
\examples{
data(table6_10)
str(table6_10)

}
\keyword{datasets}
