% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table7_01}
\alias{table7_01}
\title{Table 7.1}
\format{
A data frame with 24 rows and 3 variables.
 \describe{
  \item{specimen}{Concrete specimen tested}
  \item{concrete}{Concrete formula (1 through 8)}
  \item{strength}{28-day compressive strength (psi)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 445
}
\usage{
data(table7_01)
}
\description{
Table 7.1
}
\examples{
data(table7_01)
str(table7_01)

}
\keyword{datasets}
