% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table7_02}
\alias{table7_02}
\title{Table 7.2}
\format{
A data frame with 19 rows and 2 variables.
 \describe{
  \item{type}{Spring type (1, 2, or 3)}
  \item{k}{Measured spring constant}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 446
}
\usage{
data(table7_02)
}
\description{
Table 7.2
}
\examples{
data(table7_02)
str(table7_02)

}
\keyword{datasets}
