% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table7_08}
\alias{table7_08}
\title{Table 7.8}
\format{
A data frame with 4 rows and 3 variables.
 \describe{
  \item{brand}{Brand of paper towels}
  \item{n}{Sample size}
  \item{y_bar}{Mean absorbency}
  \item{s}{Standard deviation of absorbency}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 466
}
\usage{
data(table7_08)
}
\description{
Table 7.8
}
\examples{
data(table7_08)
str(table7_08)

}
\keyword{datasets}
