% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table7_09}
\alias{table7_09}
\title{Table 7.9}
\format{
A data frame with 12 rows and 3 variables.
 \describe{
  \item{pct_water}{Percent water in mix of bricks}
  \item{heat}{Heat treatment}
  \item{MOR}{Modulus of rupture value (psi)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 467
}
\usage{
data(table7_09)
}
\description{
Table 7.9
}
\examples{
data(table7_09)
str(table7_09)

}
\keyword{datasets}
