% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table7_14}
\alias{table7_14}
\title{Table 7.14}
\format{

}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 488
}
\usage{
data(table7_14)
}
\description{
Table 7.14
}
\examples{
data(table7_14)
str(table7_14)

}
\keyword{datasets}
