% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table7_16}
\alias{table7_16}
\title{Table 7.16}
\format{
A data frame with 110 rows and 2 variables.
 \describe{
  \item{sample}{Sample ID}
  \item{excess}{Lengtth of rolled sheet in 1/64 in. over reference length}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 488
}
\usage{
data(table7_16)
}
\description{
Table 7.16
}
\examples{
data(table7_16)
str(table7_16)

}
\keyword{datasets}
