% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table7_18}
\alias{table7_18}
\title{Table 7.18}
\format{
A data frame with 25 rows and 3 variables.
 \describe{
  \item{sample}{Sample ID}
  \item{nonconform}{Number of nonconforming pellets out of 30}
  \item{p_hat}{Proportion of nonconforming pellets out of 30}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 520
}
\usage{
data(table7_18)
}
\description{
Table 7.18
}
\examples{
data(table7_18)
str(table7_18)

}
\keyword{datasets}
