% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table7_19}
\alias{table7_19}
\title{Table 7.19}
\format{
A data frame with 30 rows and 4 variables.
 \describe{
  \item{sample}{Sample ID}
  \item{trucks}{Number of trucks produced}
  \item{errors}{Errors found}
  \item{errors_truck}{Errors per truck}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 526
}
\usage{
data(table7_19)
}
\description{
Table 7.19
}
\examples{
data(table7_19)
str(table7_19)

}
\keyword{datasets}
