% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table_a_2}
\alias{table_a_2}
\title{Table A.2}
\format{
A data frame with 23 rows and 3 variables.
 \describe{
  \item{flight_date}{Date of flight of space shuttle Challenger}
  \item{o_ring}{Number of field joint primary O-ring incidents}
  \item{tempF}{Temperature at launch (F)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 768
}
\usage{
data(table_a_2)
}
\description{
Table A.2
}
\examples{
data(table_a_2)
str(table_a_2)
}
\keyword{datasets}
