% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table_a_4}
\alias{table_a_4}
\title{Table A.4}
\format{
A data frame with 10 rows and 1 variable.
 \describe{
  \item{failure}{Failure times for hardened steel specimens subjected to rolling fatigue test}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 778
}
\usage{
data(table_a_4)
}
\description{
Table A.4
}
\examples{
data(table_a_4)
str(table_a_4)
}
\keyword{datasets}
