% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table_a_5}
\alias{table_a_5}
\title{Table A.5}
\format{
A data frame with 12 rows and 1 variable.
 \describe{
  \item{breakdown}{Time (sec) to breakdown of an insulating fluid at 30 kV}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 780
}
\usage{
data(table_a_5)
}
\description{
Table A.5
}
\note{
Largest two times are censored.
}
\examples{
data(table_a_5)
str(table_a_5)
}
\keyword{datasets}
