% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch4_eoc_prob04}
\alias{ch4_eoc_prob04}
\title{Chapter 4, Exercise 4}
\format{
A data frame with 8 rows and 4 variables.
 \describe{
  \item{A}{Initial epoxy/isocyanate ratio, coded -1 and 1}
  \item{B}{Flexibilizer concentration, coded -1 and 1}
  \item{C}{Accelerator concentration, coded -1 and 1}
  \item{y}{Impact strength, kJ/sq mm}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 205
}
\usage{
data(ch4_eoc_prob04)
}
\description{
Chapter 4, Exercise 4
}
\examples{
data(ch4_eoc_prob04)
str(ch4_eoc_prob04)

}
\keyword{datasets}
