% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{ch6_sec3_prob04}
\alias{ch6_sec3_prob04}
\title{Section 6.3, Exercise 4}
\format{
A data frame with 6 rows and 2 variableS.
 \describe{
  \item{Treaded}{Stopping distance with treaded tires (cm)}
  \item{Smooth}{Stopping distance with smooth tires (cm)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigita5press.com/plugins/books/127/}, p. 385
}
\usage{
data(ch6_sec3_prob04)
}
\description{
Section 6.3, Exercise 4
}
\note{
Data used in Section 6.4, Exercise 1
}
\examples{
data(ch6_sec3_prob04)
str(ch6_sec3_prob04)

}
\keyword{datasets}
