% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table1_01}
\alias{table1_01}
\title{Table 1.1}
\format{
A data frame with 39 rows and 2 variables.
 \describe{
  \item{Laid}{Thrust face runouts for laid gears (0.0001 in.)}
  \item{Hung}{Thrust face runouts for hung gears (0.0001 in.)}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) Basic Engineering Data Collection and Analysis
\url{https://www.iastatedigitalpress.com/plugins/books/127/}
}
\usage{
data(table1_01)
}
\description{
Table 1.1
}
\examples{
data(table1_01)
str(table1_01)

}
\keyword{datasets}
