% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDCA-package.R
\docType{data}
\name{table4_08}
\alias{table4_08}
\title{Table 4.8}
\format{
A data frame with 17 rows and 4 variables.
 \describe{
  \item{air}{Air flow}
  \item{water}{Cooling water inlet temperature}
  \item{acid}{Acid concentration}
  \item{stack}{Stack loss}
  }
}
\source{
Vardeman SB & Jobe MJ (2023) \emph{Basic Engineering Data Collection and Analysis}
\url{https://www.iastatedigitalpress.com/plugins/books/127/}, p. 150
}
\usage{
data(table4_08)
}
\description{
Table 4.8
}
\note{
Brownlee's stack loss data, presented here with 4 cases of 21 omitted. Complete data are available as \code{datasets::stackloss}. Omitted rows are 1, 3, 4, 21.
}
\examples{
data(table4_08)
str(table4_08)

}
\references{
Brownlee KA (1965) \emph{Statistical Theory and Methodology In Science and Engineering} (second edition), New York: Wiley.
}
\keyword{datasets}
